Changes 10/2/16
Decreased the amount of code of 10+ lines to 2 lines doing the same job
but even better on the step event.

Changed the inputs script to make inputs work flowlessly now. Now the
player won't keep going on a directional unless two of the buttons that
make them are being pressed e.i right+up.

Turned off color interpolation. This is much better for crisp pixel art